langgraph_mastery/requirements.txt
langgraph>=0.0.19
langchain>=0.0.335
langchain-core>=0.1.17
langchain-google-genai>=0.0.5
google-generativeai>=0.3.1
pydantic>=2.5.2
python-dotenv>=1.0.0
jupyter>=1.0.0
matplotlib>=3.7.1
numpy>=1.24.3
